CICD: Isolate loader tests in submodules
Merge remaining loader tests to the description test fixture.
