+++ tests/__init__.py
@@ -7,13 +7,11 @@
 from .test_clone_to_cache import TestCloneToCache
 from .test_clone_to_directory import TestCloneToDirectory
 from .test_descriptions import TestDescriptions
-from .test_loaders import TestLoaders
 from .test_progress_bar import TestProgressBar
 
 __all__ = [
     "TestCloneToCache",
     "TestCloneToDirectory",
     "TestDescriptions",
-    "TestLoaders",
     "TestProgressBar",
 ]
