+++ tests/loaders/test_mujoco.py
@@ -7,14 +7,18 @@
 import unittest
 
 from robot_descriptions._descriptions import DESCRIPTIONS
-from robot_descriptions.loaders.mujoco import load_robot_description
+from robot_descriptions.loaders.mujoco import (
+    load_robot_description as load_mujoco,
+)
 
 
 class TestMuJoCo(unittest.TestCase):
-
     """
     Check that all MJCF descriptions are loaded properly in MuJoCo.
     """
+
+    def test_mujoco(self):
+        self.assertIsNotNone(load_mujoco("cassie_mj_description"))
 
     @staticmethod
     def get_test_for_description(description: str):
@@ -29,7 +33,7 @@ class TestMuJoCo(unittest.TestCase):
         """
 
         def test(self):
-            load_robot_description(description)
+            load_mujoco(description)
 
         return test
