+++ tests/test_descriptions.py
@@ -12,6 +12,9 @@ from importlib import import_module  # type: ignore
 import git
 
 from robot_descriptions._descriptions import DESCRIPTIONS
+from robot_descriptions.loaders.pinocchio import (
+    load_robot_description as load_pinocchio,
+)
 
 
 class TestDescriptions(unittest.TestCase):
@@ -46,6 +49,13 @@ class TestDescriptions(unittest.TestCase):
                     f"in {description}",
                 )
 
+    def test_cache_path_package_name(self):
+        """Check a description with package:// URIs and a custom commit ID."""
+        load_pinocchio(
+            "draco3_description",
+            commit="5afd19733d7b3e9f1135ba93e0aad90ed1a24cc7",
+        )
+
     def test_invalid_description_commit(self):
         invalid_commit = "foobar"
         os.environ["ROBOT_DESCRIPTION_COMMIT"] = invalid_commit
@@ -55,3 +65,11 @@ class TestDescriptions(unittest.TestCase):
             if description_name in sys.modules:
                 del sys.modules[description_name]
             import_module(description_name)
+
+    def test_load_with_commit_then_without(self):
+        # https://github.com/robot-descriptions/robot_descriptions.py/issues/67
+        load_pinocchio(
+            "draco3_description",
+            commit="5afd19733d7b3e9f1135ba93e0aad90ed1a24cc7",
+        )
+        load_pinocchio("baxter_description")
