+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 isolated_build = True
-envlist = coverage,lint,loader-{mujoco,pinocchio,pybullet,robomeshcat,yourdfpy,idyntree}
+envlist = coverage,lint,loader-{mujoco,pybullet,robomeshcat,yourdfpy,idyntree}
 
 [gh-actions]
 python =
